+++ lib/routers.js
@@ -3,9 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.Link = exports.BaseRouter = exports.RouteUtils = undefined;
-
-var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+exports.BaseRouter = exports.RouteUtils = undefined;
 
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
@@ -25,10 +23,6 @@ var _react2 = _interopRequireDefault(_react);
 var _components = require('./components');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -310,96 +304,3 @@ var BaseRouter = exports.BaseRouter = function () {
 
   return BaseRouter;
 }();
-
-var Link = exports.Link = function (_BaseComponent) {
-  _inherits(Link, _BaseComponent);
-
-  function Link() {
-    _classCallCheck(this, Link);
-
-    return _possibleConstructorReturn(this, (Link.__proto__ || Object.getPrototypeOf(Link)).apply(this, arguments));
-  }
-
-  _createClass(Link, [{
-    key: 'handleClick',
-    value: function handleClick(event) {
-      var allowTransition = true;
-
-      if (this.props.onClick) {
-        this.props.onClick(event);
-      }
-
-      if (Link.isModifiedEvent(event) || !Link.isLeftClickEvent(event)) {
-        return;
-      }
-
-      if (event.defaultPrevented === true) {
-        allowTransition = false;
-      }
-
-      // If target prop is set (e.g. to "_blank") let browser handle link.
-      if (this.props.target) {
-        if (!allowTransition) {
-          event.preventDefault();
-        }
-        return;
-      }
-
-      event.preventDefault();
-
-      if (allowTransition) {
-        var _props = this.props;
-        var state = _props.state;
-        var params = _props.params;
-
-        this.response.navigate(state, params);
-      }
-    }
-  }, {
-    key: 'render',
-    value: function render() {
-      var props = {};
-      var _props2 = this.props;
-      var state = _props2.state;
-      var params = _props2.params;
-      var className = _props2.className;
-      var activeClassName = _props2.activeClassName;
-      var children = _props2.children;
-      var style = _props2.style;
-      var title = _props2.title;
-
-
-      props.href = RouteUtils.reverse(state, params);
-
-      if (activeClassName) {
-        if (this.request && this.request.isState(this.props.state)) {
-          props.className = className === '' ? activeClassName : className + ' ' + activeClassName;
-        }
-      }
-
-      if (style) {
-        props.style = style;
-      }
-
-      if (title) {
-        props.title = title;
-      }
-
-      props.children = children;
-
-      return _react2.default.createElement('a', _extends({}, props, { onClick: this.handleClick.bind(this) }));
-    }
-  }], [{
-    key: 'isLeftClickEvent',
-    value: function isLeftClickEvent(event) {
-      return event.button === 0;
-    }
-  }, {
-    key: 'isModifiedEvent',
-    value: function isModifiedEvent(event) {
-      return !!(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey);
-    }
-  }]);
-
-  return Link;
-}(_components.BaseComponent);
